% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_dqc.R
\name{cat_dqc}
\alias{cat_dqc}
\title{Create a string to use as a Detected-Quarantine-Community vector input}
\usage{
cat_dqc(
  Ds = 0,
  Da = 0,
  Qcds = 0,
  Qhds = 0,
  Qcda = 0,
  Qhda = 0,
  Qq = 0,
  Cs = 0.8,
  Ca = 0.2
)
}
\arguments{
\item{Ds}{Proportion of population that are symptomatic, detected and
isolated}

\item{Da}{Proportion of population that are asymptomatic, detected and
isolated}

\item{Qcds}{Proportion of population that are infected community contacts
of symptomatic, isolated infections and are quarantined}

\item{Qhds}{Proportion of population that are infected household contacts
of symptomatic, isolated infections and are quarantined}

\item{Qcda}{Proportion of population that are infected community contacts
of asymptomatic, isolated infections and are quarantined}

\item{Qhda}{Proportion of population that are infected household contacts
of symptomatic, isolated infections and are quarantined}

\item{Qq}{Proportion of population that were infected by quarantined
contacts and are quarantined}

\item{Cs}{Proportion of population that are symptomatic and undetected
in the community}

\item{Ca}{Proportion of population that are asymptomatic and undetected
in the community}
}
\value{
Character string.
}
\description{
Creates a string formatted as a named vector to use as the \code{dqc} or \code{init}
parameter in tti functions.
}
