% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_pqc.R
\name{cat_pqc}
\alias{cat_pqc}
\title{Create a string to use as a Passive-Quarantine-Community vector input}
\usage{
cat_pqc(Ps = 0, Pa = 0, Qcps = 0, Qhps = 0, Qcpa = 0, Qhpa = 0,
  Qq = 0, Cs = 0.8, Ca = 0.2)
}
\arguments{
\item{Ps}{Proportion of population that are symptomatic and passively
detected}

\item{Pa}{Proportion of population that are asymptomatic and passively
detected}

\item{Qcps}{Proportion of population that were infected by symptomatic
community contacts and are quarantined}

\item{Qhps}{Proportion of population that were infected by symptomatic
household contacts and are quarantined}

\item{Qcpa}{Proportion of population that were infected by asymptomatic
community contacts and are quarantined}

\item{Qhpa}{Proportion of population that were infected by asymptomatic
household contacts and are quarantined}

\item{Qq}{Proportion of population that were infected by quarantined
contacts and are quarantined}

\item{Cs}{Proportion of population that are symptomatic and undetected
in the community}

\item{Ca}{Proportion of population that are asymptomatic and undetected
in the community}
}
\value{
Character string.
}
\description{
Creates a string formatted as a named vector to use as the \code{pqc} or \code{init}
parameter in tti functions.
}
