% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detect_mat.R
\name{get_detect_mat}
\alias{get_detect_mat}
\title{Create detect matrix}
\usage{
get_detect_mat(
  alpha = 0.2,
  omega_c = 0.5,
  omega_h = 0.5,
  omega_q = 0.5,
  quarantine_days = Inf,
  rho_s = 0.1,
  rho_a = 0.05
)
}
\arguments{
\item{alpha}{Numeric value between 0 and 1. The probability of an asymptomatic
infection. Default: 0.2.}

\item{omega_c}{Numeric value between 0 and 1. The probability of being traced and
quarantined given infected community contact of isolated case. Default: 0.5.}

\item{omega_h}{Numeric value between 0 and 1. The probability of being traced and
quarantined given infected household contact of isolated case. Default: 0.5.}

\item{omega_q}{Numeric value between 0 and 1. The probability of being traced and
quarantined given infected contact of quarantined person. Default: 0.5.}

\item{quarantine_days}{Positive numeric value. The number of days contacts are told
to quarantine. Default: Inf.}

\item{rho_s}{Numeric value between 0 and 1. The probability of detection
and isolation given symptomatic. Default: 0.1.}

\item{rho_a}{Numeric value between 0 and 1. The probability of detection
and isolation given asymptomatic. Default: 0.05.}
}
\value{
Matrix
}
\description{
This function will create the DETECT matrix
}
