% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dqc_equilibrium.R
\name{get_dqc_equilibrium}
\alias{get_dqc_equilibrium}
\title{Calculate the Detected-Quarantine-Community vector equilibrium}
\usage{
get_dqc_equilibrium(
  init = c(Ds = 0, Da = 0, Qcds = 0, Qhds = 0, Qcda = 0, Qhda = 0, Qq = 0, Cs = 0.8, Ca
    = 0.2),
  tolerance = 1e-04,
  burn_in = 10,
  ...
)
}
\arguments{
\item{init}{Named vector of initial values indicating proportion infected in
each detected-community-quarantine category. All elements of the vector
must sum to 1. Must include the names:
\code{Ds}, \code{Da}, \code{Qcds}, \code{Qhds}, \code{Qcda}, \code{Qhda}, \code{Qq}, \code{Cs}, \code{Ca}.
Default: \code{c(Ds = 0, Da = 0, Qcds = 0, Qhds = 0, Qcda = 0, Qhda = 0, Qq = 0, Cs = 0.8, Ca = 0.2)}
See \code{\link[=init_dqc]{init_dqc()}} to generate this vector.}

\item{tolerance}{Numeric value between 0 and 1. Tolerance to determine whether an
equilibrium has been met. Default: 1e-4}

\item{burn_in}{Numeric positive value. Minimum number of iterations to run.
Default: 10.}

\item{...}{Optional parameters to pass to \code{\link[=get_infect_mat]{get_infect_mat()}} and
\code{\link[=get_detect_mat]{get_detect_mat()}}. See \code{\link[=get_infect_mat]{get_infect_mat()}} and \code{\link[=get_detect_mat]{get_detect_mat()}}
for defaults.}
}
\value{
A named vector, the Detected-Quarantine-Community proportions at
equilibrium.
}
\description{
This function iterates through the Detected-Quarantine-Community (DQC) vector
until an equilibrium is reached at a given tolerance.
}
