% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dqc_equilibrium.R
\name{get_dqc_equilibrium_base}
\alias{get_dqc_equilibrium_base}
\title{Calculate the vector equilibrium}
\usage{
get_dqc_equilibrium_base(init, infect, detect, tolerance = 1e-04, burn_in = 10)
}
\arguments{
\item{init}{Initial values for DQC compartments. All elements of this vector
should sum to 1.}

\item{infect}{Matrix. The infection matrix.}

\item{detect}{Matrix. The detection matrix.}

\item{tolerance}{Numeric value between 0 and 1. Tolerance to determine whether an
equilibrium has been met. Default: 1e-4}

\item{burn_in}{Numeric positive value. Minimum number of iterations to run.
Default: 10.}
}
\value{
The proportions at equilibrium.
}
\description{
This low-level function iterates through the initial DQC vector
until an equilibrium is reached at a given tolerance.
}
