% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dqc_stoch.R
\name{get_dqc_stoch}
\alias{get_dqc_stoch}
\title{Calculate the Detected-Quarantine-Community vector stochastically}
\usage{
get_dqc_stoch(
  init = c(Ds = 0, Da = 0, Qcds = 0, Qhds = 0, Qcda = 0, Qhda = 0, Qq = 0, Cs = 0.8, Ca
    = 0.2),
  n_generation = 10,
  ...
)
}
\arguments{
\item{init}{Named vector of initial values indicating proportion infected in
each detected-community-quarantine category. All elements of the vector
must sum to 1. Must include the names:
\code{Ds}, \code{Da}, \code{Qcds}, \code{Qhds}, \code{Qcda}, \code{Qhda}, \code{Qq}, \code{Cs}, \code{Ca}.
Default: \code{c(Ds = 0, Da = 0, Qcds = 0, Qhds = 0, Qcda = 0, Qhda = 0, Qq = 0, Cs = 0.8, Ca = 0.2)}
See \code{\link[=init_dqc]{init_dqc()}} to generate this vector.}

\item{n_generation}{Numeric positive value. Number of iterations to run.
Default: 10.}

\item{...}{Optional parameters to pass to \code{\link{get_intermediate_inf}} and
\code{\link[=get_detect_mat]{get_detect_mat()}}. See \code{\link{get_intermediate_inf}} and \code{\link[=get_detect_mat]{get_detect_mat()}}
for defaults.}
}
\value{
A named vector, the Detected-Quarantine-Community proportions after
some number of generations
}
\description{
This function iterates through the Detected-Quarantine-Community (DQC) vector
stochastically for a set number of generations.
}
