% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_infections.R
\name{get_n_infections}
\alias{get_n_infections}
\title{Calculate Number of Infections}
\usage{
get_n_infections(n_tests, prop_pos)
}
\arguments{
\item{n_tests}{Numeric positive value. Number of tests.}

\item{prop_pos}{Proportion of tests that are returned positive}
}
\value{
Numeric. Number of potential infections that can be detected with
testing
}
\description{
Calculates number of potential infections detected from testing + surveillance
}
