% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_passive_detect.R
\name{get_passive_detect}
\alias{get_passive_detect}
\title{Calculate passive detection rates for infections}
\usage{
get_passive_detect(n_tests, prop_pos, n_infect, alpha)
}
\arguments{
\item{n_tests}{Numeric positive value. Number of tests devoted to
detection of infection type of interest}

\item{prop_pos}{Numeric value between 0 and 1. Proportion of tests that
are returned positive among infection type of interest}

\item{n_infect}{Numeric positive value. Number of infections in the
population}

\item{alpha}{Numeric value between 0 and 1. Proportion of infections
that of infection type of interest}
}
\value{
Numeric.  Proportion of infections of type of interested that
are detected
}
\description{
Function to calculate proportion of infections
detected through testing + surveillance
}
