% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pqc_equilibrium.R
\name{get_pqc_equilibrium}
\alias{get_pqc_equilibrium}
\title{Calculate the Passive-Quarantine-Community vector equilibrium}
\usage{
get_pqc_equilibrium(
  init = c(Ps = 0, Pa = 0, Qcps = 0, Qhps = 0, Qcpa = 0, Qhpa = 0, Qq = 0, Cs = 0.8, Ca
    = 0.2),
  tolerance = 1e-04,
  burn_in = 10,
  ...
)
}
\arguments{
\item{init}{Named vector of initial values indicating proportion infected in
each passive-community-quarantine category. All elements of the vector
must sum to 1. Must include the names:
\code{Ps}, \code{Pa}, \code{Qcps}, \code{Qhps}, \code{Qcpa}, \code{Qhpa}, \code{Qq}, \code{Cs}, \code{Ca}.
Default: \code{c(Ps = 0, Pa = 0, Qcps = 0, Qhps = 0, Qcpa = 0, Qhpa = 0, Qq = 0, Cs = 0.8, Ca = 0.2)}
See \code{\link[=init_pqc]{init_pqc()}} to generate this vector.}

\item{tolerance}{Numeric value between 0 and 1. Tolerance to determine whether an
equilibrium has been met. Default: 1e-4}

\item{burn_in}{Numeric positive value. Minimum number of iterations to run.
Default: 10.}

\item{...}{Optional parameters to pass to \code{\link[=get_infect_mat]{get_infect_mat()}} and
\code{\link[=get_detect_mat]{get_detect_mat()}}. See \code{\link[=get_infect_mat]{get_infect_mat()}} and \code{\link[=get_detect_mat]{get_detect_mat()}}
for defaults.}
}
\value{
A named vector, the Passive-Quarantine-Community proportions at
equilibrium.
}
\description{
This function iterates through the Passive-Quarantine-Community (PQC) vector
until an equilibrium is reached at a given tolerance.
}
