% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_isolated.R
\name{get_prop_isolated}
\alias{get_prop_isolated}
\title{Calculate the proportion detected and isolated through testing}
\usage{
get_prop_isolated(dqc)
}
\arguments{
\item{dqc}{Numeric vector of Detected-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ds}, \code{Da}, \code{Qcds}, \code{Qcda}, \code{Qhds}, \code{Qhda},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_dqc]{init_dqc()}} to generate this vector.}
}
\value{
Numeric value. Proportion detected and isolated.
}
\description{
Calculates the proportion detected and isolated through testing
(i.e. outside of contact tracing) from a Detected-Quarantine-
Community vector
}
