% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_passive.R
\name{get_prop_passive}
\alias{get_prop_passive}
\title{Calculate the proportion quarantined}
\usage{
get_prop_passive(pqc)
}
\arguments{
\item{pqc}{Numeric vector of Passive-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ps}, \code{Pa}, \code{Qcps}, \code{Qcpa}, \code{Qhps}, \code{Qhpa},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_pqc]{init_pqc()}} to generate this vector.}
}
\value{
Numeric value. Proportion passively detected.
}
\description{
Calculates the proportion passively detected from a
Passive-Quarantine-Community vector
}
