% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_quarantined.R
\name{get_prop_quarantined}
\alias{get_prop_quarantined}
\title{Calculate the proportion quarantined}
\usage{
get_prop_quarantined(pqc)
}
\arguments{
\item{pqc}{Numeric vector of Passive-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ps}, \code{Pa}, \code{Qcps}, \code{Qcpa}, \code{Qhps}, \code{Qhpa},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_pqc]{init_pqc()}} to generate this vector.}
}
\value{
Numeric value. Proportion quarantined.
}
\description{
Calculates the proportion quarantined from a Passive-Quarantine-Community
vector
}
