% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proportions_df_stochastic.R
\name{get_proportions_df_stochastic}
\alias{get_proportions_df_stochastic}
\title{Create a data frame with the proportion infected over time}
\usage{
get_proportions_df_stochastic(duration, init = c(Ps = 0, Pa = 0, Qcps =
  0, Qhps = 0, Qcpa = 0, Qhpa = 0, Qq = 0, Cs = 0.8, Ca = 0.2), ...)
}
\arguments{
\item{duration}{Numeric value greater than 0. The number of iterations you
would like to examine.}

\item{init}{Named vector of initial values indicating proportion infected in
each passive-community-quarantine category. All elements of the vector
must sum to 1. Must include the names:
\code{Ps}, \code{Pa}, \code{Qcps}, \code{Qhps}, \code{Qcpa}, \code{Qhpa}, \code{Qq}, \code{Cs}, \code{Ca}.
Default: \code{c(Ps = 0, Pa = 0, Qcps = 0, Qhps = 0, Qcpa = 0, Qhpa = 0, Qq = 0, Cs = 0.8, Ca = 0.2)}
See \code{\link[=init_pqc]{init_pqc()}} to generate this vector.}

\item{...}{Optional parameters to pass to \code{\link[=get_infect_mat]{get_infect_mat()}} and
\code{\link[=get_detect_mat]{get_detect_mat()}}. See \code{\link[=get_infect_mat]{get_infect_mat()}} and \code{\link[=get_detect_mat]{get_detect_mat()}}
for defaults.}
}
\value{
data frame with three columns:
\itemize{
\item \code{t}: the time
\item \code{prop_infected}: the proportion infected
\item \code{r_effective}: the effective reproduction number at the given time point
\item \code{category}: the category
}
}
\description{
Create a data frame with the proportion infected over time
}
