% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_effective.R
\name{get_r_effective}
\alias{get_r_effective}
\title{Get R effective}
\usage{
get_r_effective(pqc, ...)
}
\arguments{
\item{pqc}{Numeric vector of Passive-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ps}, \code{Pa}, \code{Qcps}, \code{Qcpa}, \code{Qhps}, \code{Qhpa},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_pqc]{init_pqc()}} to generate this vector.}

\item{...}{Optional parameters to pass to \code{\link[=get_infect_mat]{get_infect_mat()}}.}
}
\value{
Numeric value indicating the R effective given the
Passive-Quarantine-Community proportions supplied.
}
\description{
Get R effective
}
