% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_effective_df.R
\name{get_r_effective_df}
\alias{get_r_effective_df}
\title{Create a data frame with R effective, varying parameters}
\usage{
get_r_effective_df(alpha = 0.2, R = 2.5, kappa = 0.5, eta = 0.5,
  nu = 4, t_ps = 3, t_pa = 3, t_qcs = 3, t_qca = 3, t_qhs = 3,
  t_qha = 3, t_q = 3, omega_c = 0.5, omega_h = 0.5,
  omega_q = 0.5, rho_s = 0.1, rho_a = 0.05)
}
\arguments{
\item{alpha}{Numeric value or vector of numeric values between 0 and 1.
The probability of an asymptomatic infection. Default 0.2.}

\item{R}{Positive numeric value or vector of positive numeric values.
Reproduction number. Default: 2.5.}

\item{kappa}{Positive numeric value or vector of positive numeric values. Relative transmissibility of an
asymptomatic individual compared to a symptomatic individual.
Default: 0.5.}

\item{eta}{Numeric value or vector of numeric values between 0 and 1. Probability contact is a household
contact. Default: 0.5}

\item{nu}{Positive numeric value or vector of positive numeric values. Relative risk of infection for a household
contact compared to a community contact. Default: 4.}

\item{t_ps}{Positive numeric value. Time delay from symptom onset to
isolation in passively detected symptomatic person. Default: 3.}

\item{t_pa}{Positive numeric value. Time delay from symptom onset to
isolation in passively detected asymptomatic person. Default: 3.}

\item{t_qcs}{Positive numeric value. Time delay from symptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_qca}{Positive numeric value. Time delay from asymptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_qhs}{Positive numeric value. Time delay from symptomatic index
cases's symptom onset to quarantine of household contacts.
Default: 3.}

\item{t_qha}{Positive numeric value. Time delay from asymptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_q}{Positive numeric value. Time delay from quarantined index
cases's symptom onset to quarantine of contacts.
Default: 3.}

\item{omega_c}{Numeric value or vector of numeric values between 0 and 1. The probability of being traced and
quarantined given community contact of a person. Default: 0.5.}

\item{omega_h}{Numeric value or vector of numeric values between 0 and 1. The probability of being traced and
quarantined given household contact of a person. Default: 0.5.}

\item{omega_q}{Numeric value or vector of numeric values between 0 and 1. The probability of being traced and
quarantined given quarantine contact of a person. Default: 0.5.}

\item{rho_s}{Numeric value or vector of numeric values between 0 and 1. The probability of passive
detection given symptomatic. Default: 0.1.}

\item{rho_a}{Numeric value or vector of numeric values between 0 and 1. The probability of passive
detection given asymptomatic. Default: 0.05.}
}
\value{
Data frame with 18 columns:
\itemize{
\item \code{r_effective}: The R effective value
\item \code{alpha}
\item \code{R}
\item \code{kappa}
\item \code{eta}
\item \code{nu}
\item \code{t_ps}
\item \code{t_pa}
\item \code{t_qcs}
\item \code{t_qca}
\item \code{t_qhs}
\item \code{t_qha}
\item \code{t_q}
\item \code{omega_c}
\item \code{omega_h}
\item \code{omega_q}
\item \code{rho_s}
\item \code{rho_a}
}
}
\description{
This function allows you to input numeric scalars or vectors for each
parameter to examine how the effective R will vary across configurations.
}
\examples{
get_r_effective_df(alpha = c(0.1, 0.2), kappa = c(0.5, 0.6))
}
