% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_effective_stoch.R
\name{get_r_effective_stoch}
\alias{get_r_effective_stoch}
\title{Get R effective for stochastic model}
\usage{
get_r_effective_stoch(dqc, n_inf = 1000, ...)
}
\arguments{
\item{dqc}{Numeric vector of Detected-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ds}, \code{Da}, \code{Qcds}, \code{Qcda}, \code{Qhds}, \code{Qhda},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_dqc]{init_dqc()}} to generate this vector.}

\item{n_inf}{Number of infections (or effective population size of infected
individuals). Default 1000.}

\item{...}{Optional parameters to pass to \code{\link[=get_intermediate_inf]{get_intermediate_inf()}}.}
}
\value{
Numeric value indicating the R effective given the
Detected-Quarantine-Community proportions supplied.
}
\description{
Get R effective for stochastic model
}
