% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_stochastic.R
\name{make_input_stochastic}
\alias{make_input_stochastic}
\title{Create an input dataframe that serve as parameter input for stochastic simulations}
\usage{
make_input_stochastic(nsimulations, input_list, input_vartype)
}
\arguments{
\item{nsimulations}{number of stochastic simulations}

\item{input_list}{list of input parameters with named elements: alpha, omega_c, omega_h, omega_q, rho_s, rho_a, R, kappa, eta, nu, t_ps, t_pa, t_qcs, t_qca, t_qhs, t_qha, t_q}

\item{input_vartype}{list of input parameter variable types with named elements: alpha, omega_c, omega_h, omega_q, rho_s, rho_a, R, kappa, eta, nu, t_ps, t_pa, t_qcs, t_qca, t_qhs, t_qha, t_q. Currently supported character options are "proportion", "R_overdispersed", "uniform", "poisson", "fixed"}
}
\value{
dataframe
}
\description{
Create an input dataframe that serve as parameter input for stochastic simulations
}
