% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infect_mat.R
\name{get_infect_mat}
\alias{get_infect_mat}
\title{Create infect matrix}
\usage{
get_infect_mat(alpha = 0.2, R = 2.5, kappa = 0.5, eta = 0.5,
  nu = 4, t_ps = 3, t_pa = 3, t_qcs = 3, t_qca = 3, t_qhs = 3,
  t_qha = 3, t_q = 3, t_incubation = 5.5, offset = -2.31,
  shape = 3, rate = 0.69)
}
\arguments{
\item{alpha}{Numeric value between 0 and 1. The probability of an asymptomatic
infection. Default 0.2.}

\item{R}{Positive numeric value. Reproduction number. Default: 2.5.}

\item{kappa}{Positive numeric value. Relative transmissibility of an
asymptomatic individual compared to a symptomatic individual.
Default: 0.5.}

\item{eta}{Numeric value between 0 and 1. Probability contact is a household
contact. Default: 0.5}

\item{nu}{Positive numeric value. Relative risk of infection for a household
contact compared to a community contact. Default: 4.}

\item{t_ps}{Positive numeric value. Time delay from symptom onset to
isolation in passively detected symptomatic person. Default: 3.}

\item{t_pa}{Positive numeric value. Time delay from symptom onset to
isolation in passively detected asymptomatic person. Default: 3.}

\item{t_qcs}{Positive numeric value. Time delay from symptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_qca}{Positive numeric value. Time delay from asymptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_qhs}{Positive numeric value. Time delay from symptomatic index
cases's symptom onset to quarantine of household contacts.
Default: 3.}

\item{t_qha}{Positive numeric value. Time delay from asymptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_q}{Positive numeric value. Time delay from quarantined index
cases's symptom onset to quarantine of contacts.
Default: 3.}

\item{t_incubation}{Positive numeric value. The estimated average
incubation time. Default: 5.5.}

\item{offset}{Numeric. Offset of infectiousness compared to symptoms onset.
Default is -2.31.}

\item{shape}{Numeric. Shape of the gamma distribution of infectious period.
Default is 3.}

\item{rate}{Numeric. Rate of the gamma distribution of infectious period.
Default is 0.69.}
}
\value{
Matrix
}
\description{
This function will create the INFECT matrix
}
