% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dqc_stoch.R
\name{get_intermediate_inf}
\alias{get_intermediate_inf}
\title{Calculate the intermediate infection states}
\usage{
get_intermediate_inf(dqc, infect, n_inf = 1000, theta = 0.2)
}
\arguments{
\item{dqc}{Current values for DQC compartments. All elements of this vector
should sum to 1.}

\item{infect}{INFECT matrix returned by \code{\link{get_infect_mat}}}

\item{n_inf}{Number of infections (or effective population size of infected
individuals). Default 1000.}

\item{theta}{Non-negative numeric value. Shape parameter for overdispersion
in R. Default 0.2.}
}
\value{
Named vector of intermediate infection states
}
\description{
This low-level function stochastically derives the intermediate infections derived
from a DQC matrix in one generation
}
