% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_identified.R
\name{get_prop_identified}
\alias{get_prop_identified}
\title{Calculate the proportion isolated and quarantined}
\usage{
get_prop_identified(dqc)
}
\arguments{
\item{dqc}{Numeric vector of Detected-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ds}, \code{Da}, \code{Qcds}, \code{Qcda}, \code{Qhds}, \code{Qhda},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_dqc]{init_dqc()}} to generate this vector.}
}
\value{
Numeric value. Proportion identified.
}
\description{
Calculates the proportion isolated and quarantined from a
Detected-Quarantine-Community vector
}
