% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_infect_time.R
\name{get_prop_infect_time}
\alias{get_prop_infect_time}
\title{Calculate proportion of the infectious period reduced}
\usage{
get_prop_infect_time(delay, offset = -12.27, shape = 21.13, rate = 1.592)
}
\arguments{
\item{delay}{Numeric greater than 0. Time from symptom onset to isolation}

\item{offset}{Numeric. Offset of infectiousness compared to symptom onset.
Default is -2.31.}

\item{shape}{Numeric. Shape of the gamma distribution of infectious period.
Default is 1.65.}

\item{rate}{Numeric. Rate of the gamma distribution of infectious period.
Default is 0.5.}
}
\value{
Numeric. Proportion of infectious period reduced given the
distribution of infectiousness and time from symptom onset to isolation.
}
\description{
Calculates the proportion of the infections period reduced given the
distribution of infectiousness and time from symptom onset to isolation.
Infectious period is assumed to be gamma distributed.
}
