% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_quarantined.R
\name{get_prop_quarantined}
\alias{get_prop_quarantined}
\title{Calculate the proportion quarantined}
\usage{
get_prop_quarantined(dqc)
}
\arguments{
\item{dqc}{Numeric vector of Detected-Quarantine-Community proportions.
The components of this vector should sum to one. The order of the
components should be:  \code{Ds}, \code{Da}, \code{Qcds}, \code{Qcda}, \code{Qhds}, \code{Qhda},
\code{Qq}, \code{Cs}, \code{Ca}. See \code{\link[=init_dqc]{init_dqc()}} to generate this vector.}
}
\value{
Numeric value. Proportion quarantined.
}
\description{
Calculates the proportion quarantined from a Detected-Quarantine-Community
vector
}
