% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_effective.R
\name{get_r_effective_base}
\alias{get_r_effective_base}
\title{Get R Effective}
\usage{
get_r_effective_base(dqc, infect)
}
\arguments{
\item{dqc}{Values for DQC compartments. All elements of this vector
should sum to 1.}

\item{infect}{Matrix. The infection matrix.}
}
\description{
This low-level function calculates R effective from a DQC vector and
infection matrix
}
