% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_effective_df.R
\name{get_r_effective_df}
\alias{get_r_effective_df}
\title{Create a data frame with R effective, varying parameters}
\usage{
get_r_effective_df(
  alpha = 0.2,
  R = 2.5,
  kappa = 0.5,
  eta = 0.5,
  nu = 4,
  t_ds = 3,
  t_da = 3,
  t_qcs = 3,
  t_qca = 3,
  t_qhs = 3,
  t_qha = 3,
  t_q = 3,
  omega_c = 0.5,
  omega_h = 0.5,
  omega_q = 0.5,
  quarantine_days = Inf,
  rho_s = 0.1,
  rho_a = 0.05,
  t_incubation = 5.5,
  offset = -2.31,
  shape = 1.65,
  rate = 0.5,
  stoch = FALSE,
  theta = NULL,
  n_inf = NULL,
  n_iter = NULL
)
}
\arguments{
\item{alpha}{Numeric value or vector of numeric values between 0 and 1.
The probability of an asymptomatic infection. Default 0.2.}

\item{R}{Positive numeric value or vector of positive numeric values.
Reproduction number. Default: 2.5.}

\item{kappa}{Positive numeric value or vector of positive numeric values. Relative transmissibility of an
asymptomatic individual compared to a symptomatic individual.
Default: 0.5.}

\item{eta}{Numeric value or vector of numeric values between 0 and 1. Probability contact is a household
contact. Default: 0.5}

\item{nu}{Positive numeric value or vector of positive numeric values. Relative risk of infection for a household
contact compared to a community contact. Default: 4.}

\item{t_ds}{Non-negative numeric value. Time delay from symptom onset to
isolation in detected symptomatic person. Default: 3.}

\item{t_da}{Non-negative numeric value. Time delay from symptom onset to
isolation in detected asymptomatic person. Default: 3.}

\item{t_qcs}{Non-negative numeric value. Time delay from symptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_qca}{Non-negative numeric value. Time delay from asymptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_qhs}{Non-negative numeric value. Time delay from symptomatic index
cases's symptom onset to quarantine of household contacts.
Default: 3.}

\item{t_qha}{Non-negative numeric value. Time delay from asymptomatic index
cases's symptom onset to quarantine of community contacts.
Default: 3.}

\item{t_q}{Non-negative numeric value. Time delay from quarantined index
cases's symptom onset to quarantine of contacts.
Default: 3.}

\item{omega_c}{Numeric value or vector of numeric values between 0 and 1. The probability of being traced and
quarantined given community contact of a person. Default: 0.5.}

\item{omega_h}{Numeric value or vector of numeric values between 0 and 1. The probability of being traced and
quarantined given household contact of a person. Default: 0.5.}

\item{omega_q}{Numeric value or vector of numeric values between 0 and 1. The probability of being traced and
quarantined given quarantine contact of a person. Default: 0.5.}

\item{quarantine_days}{Positive numeric value. The number of days contacts are told
to quarantine. Default: 14.}

\item{rho_s}{Numeric value or vector of numeric values between 0 and 1. The probability of
detection and isolation given symptomatic. Default: 0.1.}

\item{rho_a}{Numeric value or vector of numeric values between 0 and 1. The probability of
detection and isolation given asymptomatic. Default: 0.05.}

\item{t_incubation}{Non-negative numeric value. The estimated average
incubation time. Default: 5.5.}

\item{offset}{Numeric. Offset of infectiousness compared to symptoms onset.
Default is -2.31.}

\item{shape}{Numeric. Shape of the gamma distribution of infectious period.
Default is 1.65.}

\item{rate}{Numeric. Rate of the gamma distribution of infectious period.
Default is 0.5.}

\item{stoch}{Logical. Whether to run stochastic model with overdispersion.
Default is FALSE.}

\item{theta}{Non-negative numeric. Required only if \code{stoch}=TRUE. Overdispersion
parameter for negative binomial distribution. Default NULL.}

\item{n_inf}{Non-negative numeric. Required only if \code{stoch}=TRUE. Number of
infections (or effective population size of infected individuals).
Default NULL.}

\item{n_iter}{Non-negative numeric. Required only if \code{stoch}=TRUE. Number of
iterations of stochastic model to run for each unique parameter value.
Default NULL.}
}
\value{
Data frame with columns:
\itemize{
\item \code{r_effective}: The R effective value
\item \code{prop_identified}: The proportion of infections identified
\item \code{alpha}
\item \code{R}
\item \code{kappa}
\item \code{eta}
\item \code{nu}
\item \code{t_ds}
\item \code{t_da}
\item \code{t_qcs}
\item \code{t_qca}
\item \code{t_qhs}
\item \code{t_qha}
\item \code{t_q}
\item \code{omega_c}
\item \code{omega_h}
\item \code{omega_q}
\item \code{quarantine_days}
\item \code{rho_s}
\item \code{rho_a}
\item \code{theta} if \code{stoch}=TRUE
\item \code{n_inf} if \code{stoch}=TRUE
}
}
\description{
This function allows you to input numeric scalars or vectors for each
parameter to examine how the effective R will vary across configurations.
Two values (\code{stoch}, \code{n_iter}) require single values.
}
\examples{
get_r_effective_df(alpha = c(0.1, 0.2), kappa = c(0.5, 0.6))
get_r_effective_df(stoch = TRUE, theta = c(0.1, 0.3), n_inf = 100, n_iter = 100)
}
